refactor model package in the test folder to move it to domain
