+++ src/test/java/uk/gov/hmcts/reform/wataskmonitor/domain/camunda/CamundaTaskTest.java
@@ -1,16 +1,15 @@
-package uk.gov.hmcts.reform.wataskmonitor.models.caseeventhandler;
+package uk.gov.hmcts.reform.wataskmonitor.domain.camunda;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wataskmonitor.domain.caseeventhandler.EventInformation;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
 
-class EventInformationTest {
+class CamundaTaskTest {
 
     @Test
     void isWellImplemented() {
-        final Class<?> classUnderTest = EventInformation.class;
+        final Class<?> classUnderTest = CamundaTask.class;
         assertPojoMethodsFor(classUnderTest)
             .testing(Method.GETTER)
             .testing(Method.CONSTRUCTOR)
