--- src/test/java/uk/gov/hmcts/reform/wataskmonitor/services/jobs/adhoc/createtasks/JsonResourceCaseRetrieverServiceTest.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.reform.wataskmonitor.services.jobs.adhoc.createtasks;
-
-import org.junit.jupiter.api.Test;
-
-class JsonResourceCaseRetrieverServiceTest {
-
-    @Test
-    void retrieveCaseList() {
-    }
-}
