--- src/test/resources/uk/gov/hmcts/reform/wataskmonitor/models/jobs/adhoc/createtasks/ad-hoc-create-tasks.json
@@ -1,6 +1,0 @@
-{
-  "caseIds": [
-    "1626272789070361",
-    "1626272789070362"
-  ]
-}
